# pos-order/src/main/resources/api.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Order
  license:
    name: MIT
paths:
  /order:
    post:
      summary: Create an order
      operationId: createOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        '200':
          description: Order created
          content:
            application/json:
              schema:
                type: integer
                format: int64
        '400':
          description: Invalid order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /order/{id}:
    get:
      summary: Get an order
      operationId: getOrder
      parameters:
        - name: id
          in: path
          description: Order ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Order found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /order/all:
    get:
      summary: Get all orders
      operationId: getAllOrders
      responses:
        '200':
          description: Orders found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '404':
          description: Orders not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Product:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
        name:
          type: string
        price:
          type: double
        image:
          type: string
    Products:
      type: array
      items:
        $ref: "#/components/schemas/Product"
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
    Item:
      type: object
      required:
        - product
        - quantity
      properties:
        product:
          $ref: "#/components/schemas/Product"
        quantity:
          type: integer
          format: int32
    Cart:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Item"
    Order:
      type: object
      required:
        - cart
        - total
      properties:
        id:
          type: integer
          format: int64
        cart:
          $ref: "#/components/schemas/Cart"
        total:
          type: number